backend/requirements.txt
# Python dependencies for Lambda functions
pymysql==1.1.0
boto3==1.34.0
python-dotenv==1.0.0
requests==2.31.0
reportlab==4.0.7
pdfrw==0.4
Pillow==10.1.0
